Fix EtagCache ENV serialization/caching
The EtagCache middleware caches the response `env` structure to allow restoring a cached response, but the `body` attribute is no longer real data, it dynamically dispatches to either `request_body` or `response_body`.  This causes an `env` being cached with a `response_body` value but no `body` value on cache hit/read, triggering a nil body to downstream middleware.