Replace mime-types dependency with mini_mime
The mime-types gem is used to determine the MIME type of the uploaded
file via the filename.

The [mini_mime](https://github.com/discourse/mini_mime) library is an
alternative which has a lower memory footprint. [This pull request on
the mail library](https://github.com/mikel/mail/pull/1059) has some
discussion on the topic.